Support ESLint flat configuration
This commit refactors `./src/index.js` to support ESLint's "new" flat
configuration file [1] in a non-disruptive, backwards-compatible manner.

The introduction of the flat configuration changed the "shape" of the
object that plugins should export. As a result, plugin authors are
adding confis with names like `flat/recommended` (as I've done here) for
consumers ot use. Rules for the "flat"-prefixed configurations are the
same as the legacy config, but with the exported objects conforming to
the new "flat" syntax.

I followed the recommendations in the plugin migration guide [2] with
inspiration drawn from the eslint-plugin-jsdoc's project [3].

Notable changes:

- adds the `meta` key as recommended by ESLint with name and version
  drawn from `package.json`
- adds a `flat/recommended` configuration matching the rules set by the
  existing `recommended configuration`

Usage in an `eslint.config.js` file would look like:

```js
import sortClassMembers from "eslint-plugin-sort-class-members";

export default [
  sortClassMembers.configs["flat/recommended"],

  rules: {
    // etc. etc. etc.
  },
];
```

[1] https://eslint.org/docs/latest/use/configure/configuration-files-new
[2] https://eslint.org/docs/latest/extend/plugin-migration-flat-config
[3] https://github.com/gajus/eslint-plugin-jsdoc/blob/main/src/index.js